Cap03 - Inserindo mensagem de confirmação de delete

+++ PSI/Models/Fabricante.cs
@@ -7,7 +7,7 @@ namespace PSI.Models
 {
     public class Fabricante
     {
-        public long FabricanteId { get; set; }
+        public int FabricanteId { get; set; }
         public string Nome { get; set; }
 
     }
